Lots of scaffolding put in place.
Copied the development version of PYSWF into the repo (the current
released version doesn't work in Python 3). Of course, this module
has some dependencies, so...

Used code from Robert Gützkow's install_dependencies example to provide
a means of installing the requisite Python modules (lxml and pylzma).

As for actually interacting with SWF files, the playground has a dummy
operator that outputs the structure of a test SWF file.