Stupid little test to import SWF by way of SVG.
The PySWF module has a built-in SVG exporter. Blender has multiple means
of importing SVGs, one as a curve and one as a grease pencil object.

Technically both work, but the preferred import should be Grease Pencil.
Unfortunately, something about PySWF's SVG export is janky because the
SVG as Grease Pencil import results in an empty GP object. The generated
SVG is valid, though... it opens in Inkscape just fine.

Maybe the GP importer is expecting something that's only in more modern
SVGs...