Made a simplifying assumption WRT coordinates.
Basically, SWF uses image-based coordinates where the origin is at the
top left of the image, X values increase to the right, and Y values
increase going down.

To make things easier on me, I've made the assumption that all drawing
happens in the lower right quadrant of the X-Y plane... this way the
only coordinate transform that's necessary is inverting imported Y
values.