Fixed some bad assumptions about transforms.
Rather than try to manually decompose the transform matrix in the SWF
file (something I was doing very poorly), I constructed a mathutils
Matrix object and used that to set the transform directly. Worked *way*
better (which is to say, it actually worked). Big thanks to Jon Denning
for pointing me in the right direction here.

Also, I think that SWF assumes that all animations are loops, so I added
a somewhat hacky piece in there at iterates through all fcurves in all
actions and adds a Cyclic modifier.