Added rudimentary support for linear gradient and bitmap fills.
Still not exactly sure what coordinate space the gradient and bitmap
matrices are in, so those offsets and angles are all wrong.

Also, the Grease Pencil HSV modifier doesn't affect image data, so it
can't be used to put a color transform on the whole GP object like in
SWF. Currently working around by playing with the layer's tint color and
mixing it. Not exact, but close.