Placeholder commit. Trying to get a handle on GP UV space.
At this point, I can reliably get the center of a radial gradient to be
at the first point in a GP stroke and I think I have the scale of the
gradient calculated correctly. The challenge is translating the UV
origin to match the center of the stroke's bounding box. There's some
coordinate space transform magic that's currently eluding me.