Found a nicer internal function in pyswf for generating paths.
This still doesn't fix my z-ordering problem, which I've determined
comes from the fact that SWF supports two different fill styles per
edge and most other applications only have a single fill style meant for
the interior space of an edge. The PYSWF library tries to handle this by
drawing that same stroke twice, but in the reverse direction for
FillStyle0. That doesn't really solve the problem as it just creates two
strokes in the same place with different materials... and the last one
drawn gets priority in terms of z-order.