MASSIVE refactor of drawing and material code
In the last [working] version of this, everything was pretty nice, but
it was all done with object animation. No effort was made to take
advantage of Grease Pencil layers. This was bad... for a number of
reasons. Z-order was one, but also any animator who'd want to work on
this imported data would likely rip their own hair out trying to manage
all of the objects and animating their visibility in the outliner. Using
layers and Grease Pencil's native support for completely changing shapes
on frame change is a much nicer way to do it.

As a bonus (and a necessity), this also meant refactoring how materials
are assigned. Now we have far fewer duplicate materials (hopefully none)
than before. Again, this should make things much easier to manage.

Of course, there are still some corner case things to finish
refactoring. Sprites are currently broken... and gradient/texture fills
remain a headache without a proper patch to Blender itself. Also, there
are some interesting line discontinuity things in the DefineShape4 test
file that aren't currently working.

But... progress!