Sprite support is back!
There's still an outstanding question of what happens if a PlaceObject
tag shows up after a sprite is placed. Right now, we're working on the
assumption that a whole new SWF object is created, but that may be wrong
and we're supposed to go back to the original SWF object.

Need more test files!