Getting closer on fixing the z-order issue.
With @vxlcoder's help, we're a little closer to getting the z-order
issue correct. Now it actually *is* a z-order issue. Before, the PYSWF
library was generating two edge sequences... one for each fill style.
That, of course, is incorrect.

Now, the add-on is creating the correct geometry with the correct
materials (holes notwithstanding). The problem, though, is that some of
those shapes are ordered incorrectly in the list of strokes. I suspect
that this is related to the reversing with new fill weirdness that PYSWF
does when it sees fillstyle_idx0 is non-zero. Not sure how to fix that,
though. Hopefully it's something easy.